+++ requestengine/serve_contracts.go
@@ -1,7 +1,6 @@
 package requestengine
 
 import (
-	"fmt"
 	"strings"
 
 	"canto-api/config"
@@ -9,25 +8,6 @@ import (
 
 	"github.com/gofiber/fiber/v2"
 )
-
-// GetGeneralContractRoutes returns a slice of routes for general contracts
-func GetGeneralContractRoutes() []string {
-	routes := []string{}
-	for _, contract := range config.ContractCalls {
-		for index, method := range contract.Methods {
-			// check if the contract has keys
-			if len(contract.Keys) == 0 {
-				// generate route from name, method and argument of contracts
-				route := contract.Name + "/" + strings.Split(method, "(")[0]
-				if len(contract.Args[index]) != 0 {
-					route += "/" + fmt.Sprintf("%v", contract.Args[index][0])
-				}
-				routes = append(routes, route)
-			}
-		}
-	}
-	return routes
-}
 
 func GetGeneralContractDataFiber(ctx *fiber.Ctx) error {
 	// assemble key from route
